Testing for upload to clouindary is good

+++ test/api/controllers/vehicle.js
@@ -148,15 +148,21 @@ exports.add_new_vehicle = async (req, res, next) => {
      } else if (validator.equals(process.env.NODE_ENV, utils.DEVELOPMENT_CLOUDINARY)) {
        var cloudinaryRename;
        for (var i = 0; i < req.files.length; i++) {
-         cloudinaryRename= await cloudinary.v2.uploader.rename(req.files[i].public_id, `test/users/${user._id}/${vehicleSaved._id}`);
+         cloudinaryRename= await cloudinary.v2.uploader.rename(req.files[i].public_id, `test/users/${user._id}/${vehicleSaved._id}/${req.files[i].public_id.split('/')[2]}.${req.files[i].format}`);
+         if (!cloudinaryRename) {
+           errorUtils.storeError(500, utils.CLOUDINARY_UPLOAD_FAIL);
+           return res.status(500).json(errorUtils.error_message(utils.CLOUDINARY_UPLOAD_FAIL, 500));
+         }
        }
      }
-
 
      res.json({ message: utils.VEHICLE_CREATED_SUCCESSFULLY });
 
    } catch (e) {
-     utils.deleteFiles(req.files);
+     if (validator.equals(process.env.NODE_ENV, utils.DEVELOPMENT)) {
+       utils.deleteFiles(req.files);
+     }
+     errorUtils.storeError(500, e.message);
      return res.status(500).json({error: e.message});
    }
 }
@@ -256,9 +262,6 @@ exports.update_vehicle = async (req, res, next) => {
     updatedVehicle = await VehicleModel.findOneAndUpdate({ _id: vehicleId, 'Dealership': userId },
                             { $inc: { totalPhotos: req.files.length } }, updateData)
                             .populate('Dealership');
-
-    console.log(updatedVehicle);
-    console.log(`req.files.length: ${req.files.length}`);
 
     if (!updatedVehicle) {
       if (includesFiles) {
